Remove XState implementation
