+++ package.json
@@ -10,21 +10,25 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^2.2.1",
+    "@stylexjs/stylex": "^0.13.1",
+    "axios": "^1.6.8",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "@reduxjs/toolkit": "^2.2.1",
-    "react-redux": "^9.1.0",
-    "axios": "^1.6.8"
+    "react-redux": "^9.1.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
     "@types/react": "^19.1.2",
     "@types/react-dom": "^19.1.2",
     "@vitejs/plugin-react": "^4.4.1",
+    "autoprefixer": "^10.4.21",
     "eslint": "^9.25.0",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "globals": "^16.0.0",
+    "postcss": "^8.5.6",
+    "tailwindcss": "^4.1.10",
     "typescript": "~5.8.3",
     "typescript-eslint": "^8.30.1",
     "vite": "^6.3.5"
